+++ packages/ai/src/generate-text/create-execute-tools-transformation.test.ts
@@ -8,7 +8,7 @@ import { beforeEach, describe, expect, it, vi } from 'vitest';
 import { z } from 'zod/v4';
 import { TypeValidationError } from '../error';
 import { asLanguageModelUsage } from '../types/usage';
-import { ModelCallEndEvent } from './core-events';
+import { ModelCallEndEvent } from './language-model-events';
 import { createExecuteToolsTransformation } from './create-execute-tools-transformation';
 import { LanguageModelStreamPart } from './stream-language-model-call';
 import {
