+++ packages/ai/src/generate-text/generate-text.ts
@@ -61,12 +61,14 @@ import { collectToolApprovals } from './collect-tool-approvals';
 import { ContentPart } from './content-part';
 import type {
   OnFinishEvent,
-  OnModelCallEndCallback,
-  OnModelCallStartCallback,
   OnStartEvent,
   OnStepFinishEvent,
   OnStepStartEvent,
 } from './core-events';
+import type {
+  OnModelCallEndCallback,
+  OnModelCallStartCallback,
+} from './language-model-events';
 import { executeToolCall } from './execute-tool-call';
 import { filterActiveTools } from './filter-active-tool';
 import { GenerateTextResult } from './generate-text-result';
